pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "codecore"
version = "0.1.0a0"
description = "Provider-agnostic terminal operating system for software development agents."
readme = "README.md"
requires-python = ">=3.11"
authors = [
  { name = "CoreEuler" }
]
keywords = ["llm", "agent", "cli", "developer-tools", "codecore"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Environment :: Console",
  "Intended Audience :: Developers",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Software Development :: Build Tools",
  "Topic :: Software Development :: Libraries :: Application Frameworks"
]
dependencies = [
  "litellm>=1.82,<2",
  "pydantic>=2.12,<3",
  "prompt_toolkit>=3.0,<4",
  "PyYAML>=6.0,<7",
  "rich>=14.0,<15",
]

[project.optional-dependencies]
dev = [
  "build>=1.4,<2",
]

[project.scripts]
codecore = "codecore.__main__:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.unittest]
start-directory = "tests"
